+++ packages/nimble-blazor/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Fri, 10 Mar 2023 22:17:50 GMT and should not be manually modified.
+This log was last generated on Tue, 28 Mar 2023 22:52:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 11.5.0
+
+Tue, 28 Mar 2023 22:52:47 GMT
+
+### Minor changes
+
+- Blazor support for anchor menu item component ([ni/nimble@fcc2f9c](https://github.com/ni/nimble/commit/fcc2f9c0a871020f8da04213e561c0953a71d62c))
+
+## 11.4.0
+
+Thu, 16 Mar 2023 22:00:14 GMT
+
+### Minor changes
+
+- Add missing properties to NimbleAnchorTab ([ni/nimble@e1746f2](https://github.com/ni/nimble/commit/e1746f201eb7919b8e44ab53caa7eba95f036ffd))
 
 ## 11.3.3